requirements.txt
python-telegram-bot==22.6
httpx==0.28.1
feedparser==6.0.12
beautifulsoup4==4.14.3
lxml==6.0.2
PyYAML==6.0.2
python-dateutil==2.9.0.post0
aiosqlite==0.22.1
tenacity==9.1.4
python-dotenv==1.0.1

pytest==8.3.5
pytest-asyncio==0.25.3
